feat(web)!: split schedule into view/edit
`/<slug>/schedule` is now the read-only view (Review section only,
visible to everyone). Editing moved to `/<slug>/schedule/edit` —
admin-only, hosting the Prelim/Elim/Skeleton tabs. Non-admins who
hit the edit URL are redirected to the view.

The view page has an Edit button for admins; the edit page has a
View link back. Both share `useScheduleData`.
