+++ apps/web/src/views/ScheduleEditView.vue
@@ -1,0 +1,283 @@
+<script setup lang="ts">
+import { computed, onMounted, ref, toRef, watch } from 'vue'
+import { useRoute, useRouter } from 'vue-router'
+
+import { estimateLaneQuizzes } from '../brackets'
+import ElimSetupSection from '../components/schedule/ElimSetupSection.vue'
+import PrelimSetupSection from '../components/schedule/PrelimSetupSection.vue'
+import SkeletonSection from '../components/schedule/SkeletonSection.vue'
+import { useScheduleData } from '../composables/useScheduleData'
+
+const props = defineProps<{ slug: string }>()
+const router = useRouter()
+const route = useRoute()
+
+const {
+  rooms,
+  slots,
+  teamCounts,
+  extraLanes,
+  loading,
+  error,
+  meet,
+  divisions,
+  isAdmin,
+  load,
+  toggleLane,
+  resizeLane,
+  createSlot,
+  updateSlot,
+  deleteSlot,
+} = useScheduleData(toRef(props, 'slug'))
+
+async function onCreateSlot(payload: {
+  startAt: string
+  durationMinutes: number
+  kind: 'quiz' | 'event'
+  eventLabel: string | null
+  sortOrder: number
+}) {
+  try {
+    await createSlot(payload)
+  } catch (e) {
+    alert((e as Error).message)
+  }
+}
+
+async function onUpdateSlot(payload: {
+  slotId: number
+  patch: {
+    startAt?: string
+    durationMinutes?: number
+    eventLabel?: string | null
+  }
+}) {
+  try {
+    await updateSlot(payload.slotId, payload.patch)
+  } catch (e) {
+    alert((e as Error).message)
+  }
+}
+
+async function onDeleteSlot(slotId: number) {
+  try {
+    await deleteSlot(slotId)
+  } catch (e) {
+    alert((e as Error).message)
+  }
+}
+
+type TabId = 'prelim' | 'elim' | 'skeleton'
+const TABS: { id: TabId; label: string }[] = [
+  { id: 'prelim', label: 'Prelim setup' },
+  { id: 'elim', label: 'Elim setup' },
+  { id: 'skeleton', label: 'Skeleton' },
+]
+const VALID_TABS = new Set<TabId>(TABS.map((t) => t.id))
+
+function tabFromHash(hash: string): TabId {
+  const id = hash.replace(/^#/, '')
+  return VALID_TABS.has(id as TabId) ? (id as TabId) : 'skeleton'
+}
+
+const activeTab = ref<TabId>(tabFromHash(route.hash))
+
+watch(
+  () => route.hash,
+  (h) => {
+    activeTab.value = tabFromHash(h)
+  },
+)
+
+function selectTab(tab: TabId) {
+  activeTab.value = tab
+  router.replace({ ...route, hash: `#${tab}` })
+}
+
+/** Total quiz budget across all divisions. Threaded down so SkeletonSection
+ *  can show capacity vs budget without recomputing. */
+const quizBudget = computed(() => {
+  let total = 0
+  for (const d of divisions.value) {
+    const teams = teamCounts.value[d] ?? 0
+    const extras = extraLanes.value[d] ?? []
+    const usedByExtras = extras.reduce((sum, l) => sum + l.teamCount, 0)
+    const mainSize = Math.max(0, teams - usedByExtras)
+    total += teams // prelim count = team count
+    total += estimateLaneQuizzes(mainSize)
+    for (const lane of extras) {
+      total += estimateLaneQuizzes(lane.teamCount)
+    }
+  }
+  return total
+})
+
+onMounted(async () => {
+  await load()
+  // Editing is admin-only; coaches/viewers fall back to the read-only
+  // view at /:slug/schedule.
+  if (!loading.value && !isAdmin.value) {
+    router.replace({ name: 'meet-schedule', params: { slug: props.slug } })
+  }
+})
+</script>
+
+<template>
+  <div class="container">
+    <button class="back-link" @click="router.push({ name: 'meet-schedule', params: { slug } })">
+      ← View schedule
+    </button>
+
+    <p v-if="loading" class="state-msg">Loading…</p>
+    <p v-else-if="error" class="state-msg state-msg--error">{{ error }}</p>
+
+    <template v-else-if="meet">
+      <div class="page-header">
+        <div class="meet-info">
+          <h2 class="page-title">Edit schedule — {{ meet.name }}</h2>
+          <span class="meet-meta">
+            {{ meet.dateFrom }}{{ meet.dateTo ? ` – ${meet.dateTo}` : '' }} ·
+            {{ divisions.length }} divisions
+          </span>
+        </div>
+      </div>
+
+      <nav class="tabs no-print" role="tablist" aria-label="Schedule sections">
+        <button
+          v-for="tab in TABS"
+          :key="tab.id"
+          type="button"
+          role="tab"
+          class="tab"
+          :class="{ 'is-active': activeTab === tab.id }"
+          :aria-selected="activeTab === tab.id"
+          @click="selectTab(tab.id)"
+        >
+          {{ tab.label }}
+        </button>
+      </nav>
+
+      <PrelimSetupSection
+        v-if="activeTab === 'prelim'"
+        :divisions="divisions"
+        :team-counts="teamCounts"
+        :editable="isAdmin"
+      />
+
+      <ElimSetupSection
+        v-else-if="activeTab === 'elim'"
+        :divisions="divisions"
+        :team-counts="teamCounts"
+        :extra-lanes="extraLanes"
+        :editable="isAdmin"
+        @toggle-lane="toggleLane"
+        @resize-lane="resizeLane"
+      />
+
+      <SkeletonSection
+        v-else-if="activeTab === 'skeleton'"
+        :rooms="rooms"
+        :slots="slots"
+        :quiz-budget="quizBudget"
+        :editable="isAdmin"
+        @create-slot="onCreateSlot"
+        @update-slot="onUpdateSlot"
+        @delete-slot="onDeleteSlot"
+      />
+    </template>
+  </div>
+</template>
+
+<style scoped>
+.container {
+  max-width: 64rem;
+  margin: 0 auto;
+  padding: 2rem 1.5rem;
+}
+
+.back-link {
+  background: none;
+  border: none;
+  padding: 0;
+  font-size: 0.8rem;
+  color: var(--color-text-faint);
+  cursor: pointer;
+  font-family: inherit;
+  margin-bottom: 1.5rem;
+  display: inline-block;
+}
+
+.back-link:hover {
+  color: var(--color-text-muted);
+}
+
+.state-msg {
+  font-size: 0.875rem;
+  color: var(--color-text-faint);
+}
+
+.state-msg--error {
+  color: var(--palette-error);
+}
+
+.page-header {
+  display: flex;
+  align-items: flex-start;
+  justify-content: space-between;
+  gap: 1rem;
+  margin-bottom: 1.25rem;
+}
+
+.meet-info {
+  display: flex;
+  flex-direction: column;
+  gap: 0.25rem;
+  flex: 1;
+  min-width: 0;
+}
+
+.page-title {
+  font-size: 1.1rem;
+  font-weight: 700;
+  color: var(--color-heading);
+}
+
+.meet-meta {
+  font-size: 0.8rem;
+  color: var(--color-text-faint);
+}
+
+.tabs {
+  display: flex;
+  gap: 0.25rem;
+  border-bottom: 1px solid var(--color-border-alt);
+  margin-bottom: 1.25rem;
+  overflow-x: auto;
+}
+
+.tab {
+  background: none;
+  border: none;
+  font: inherit;
+  font-size: 0.8rem;
+  font-weight: 600;
+  color: var(--color-text-faint);
+  cursor: pointer;
+  padding: 0.55rem 0.85rem;
+  margin-bottom: -1px;
+  border-bottom: 2px solid transparent;
+  white-space: nowrap;
+  transition:
+    color 120ms ease,
+    border-color 120ms ease;
+}
+
+.tab:hover {
+  color: var(--color-text-muted);
+}
+
+.tab.is-active {
+  color: var(--color-text);
+  border-bottom-color: var(--color-accent);
+}
+</style>
